Add Util public class and refactor Java classes
