+++ modules/augmented-reality/src/main/java/com/gluonhq/attach/ar/impl/DefaultAugmentedRealityService.java
@@ -1,20 +1,16 @@
 package com.gluonhq.attach.ar.impl;
 
-//import java.util.logging.Logger;
-
 import com.gluonhq.attach.ar.AugmentedRealityService;
-import com.gluonhq.attach.util.Constants;
+import com.gluonhq.attach.util.Util;
+
 import java.util.logging.Logger;
 
 public abstract class DefaultAugmentedRealityService implements AugmentedRealityService {
 
     private static final Logger LOG = Logger.getLogger(DefaultAugmentedRealityService.class.getName());
-    protected static boolean debug;
+    protected static boolean debug = Util.DEBUG;
     
 
     public DefaultAugmentedRealityService() {
-        if ("true".equals(System.getProperty(Constants.ATTACH_DEBUG))) {
-            debug = true;
-        }
     }
 }
